Apply workaround for RoR bug.
RoR does not tolerate '*/*' being in the ACCEPT headers. It will try to
force a html format return instead of a json format. As we only want
specified returns for json, not html in the rails backend, this causes
500, internal server error.

The only ways to fix this are to somehow get Axios to stop sending
'application/json, */*' (totally valid input btw), or get RoR to treat
everything as json.

Couldn't do the former, so I did the latter. I hate debugging magic.