Ensure that amount is a nullable number.
FormTextInput can be allowed to return "", which rather than treating as
invalid, should be maintained as 0.

We should not require users to type in 0 in this situation, and it is
also not helpful to create a transformer to ensure this is the case. If
this is not appropriate, consider creating a FormNumericTextInput
component instead of modifyiing FormTextInput.